Db and settings moved from root package to infrastructure
